requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-CORS==3.0.10
Pillow==8.3.1
python-dotenv==0.19.0
Werkzeug==2.0.2
SQLAlchemy==1.4.23
Flask-WTF==1.2.2